defined player class logic
